requirements.txt
llama-cpp-python>=0.2.0
requests>=2.25.0
flask>=2.0.0
flask-socketio>=5.0.0